+++ lib/passenger_features/home_feature/presentation/view/components/offer_box.dart
